Move AuthorDate to generic publishing component
