+++ src/components/publishing/header/__test__/header.test.tsx
@@ -6,7 +6,6 @@ import * as _ from "lodash"
 
 import { ClassicArticle, FeatureArticle, StandardArticle } from "../../fixtures/articles"
 import { HeroSections } from "../../fixtures/components"
-import AuthorDate from "../author_date"
 import AuthorDateClassic from "../author_date_classic"
 import Header from "../header"
 
@@ -36,24 +35,6 @@ describe("Standard Header", () => {
   it("renders standard header properly", () => {
     const header = renderer.create(<Header article={StandardArticle} />).toJSON()
     expect(header).toMatchSnapshot()
-  })
-})
-
-describe("AuthorDate", () => {
-  it("renders a single author", () => {
-    const authors = [{ name: "Molly Gottschalk" }]
-    const authorDate = renderer.create(
-      <AuthorDate authors={authors} date={"2017-05-19T13:09:18.567Z"} layout={"split"} />
-    )
-    expect(authorDate).toMatchSnapshot()
-  })
-
-  it("renders multiple authors", () => {
-    const authors = [{ name: "Molly Gottschalk" }, { name: "Kana Abe" }]
-    const authorDate = renderer.create(
-      <AuthorDate authors={authors} date={"2017-05-19T13:09:18.567Z"} layout={"split"} />
-    )
-    expect(authorDate).toMatchSnapshot()
   })
 })
