+++ src/components/publishing/related_articles/related_article_figure.tsx
@@ -2,6 +2,7 @@ import React from "react"
 import styled from "styled-components"
 import { crop } from "../../../utils/resizer"
 import { pMedia } from "../../helpers"
+import AuthorDate from "../author_date"
 import { articleHref } from "../constants"
 import Fonts from "../fonts"
 
@@ -10,6 +11,8 @@ interface RelatedArticleFigureProps extends React.HTMLProps<HTMLDivElement> {
     thumbnail_title: string
     thumbnail_image: string
     slug: string
+    contributing_authors: any
+    published_at: string
   }
 }
 
@@ -19,6 +22,7 @@ const RelatedArticleFigure: React.SFC<RelatedArticleFigureProps> = props => {
     <ArticleFigure href={articleHref(article.slug)}>
       <BlockImage src={crop(article.thumbnail_image, { width: 510, height: 340 })} alt={article.thumbnail_title} />
       <ArticleTitle>{article.thumbnail_title}</ArticleTitle>
+      <AuthorDate authors={article.contributing_authors} date={article.published_at} />
     </ArticleFigure>
   )
 }
